tasks/booking-history-service/src/requirements.txt
psycopg[binary]==3.2.10
kafka-python==2.2.15